exec. Since the lower half is unmapped in exec, the test for user mode in init causes a page fault.

+++ kernel/syscall_def.h
@@ -6,4 +6,4 @@ int64_t sys_write(int16_t fd, const void *buf, uint16_t count);
 
 int64_t sys_mmap(void* addr, size_t length, int prot, int flags, int fd, uint16_t offset);
 
-int sys_exec(char* name);
+int64_t sys_exec(char* name);
